+++ src/main/java/com/acltabontabon/launchpad/standards/AdapterFile.java
@@ -1,0 +1,6 @@
+package com.acltabontabon.launchpad.standards;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+record AdapterFile(int version, Adapter adapter) {}